requirements.txt
scipy==1.5.3
nibabel==3.1.0
torch==1.4.0
statsmodels==0.11.1
scikit_image==0.16.2
torchvision==0.5.0
matplotlib==3.1.3
torchsummary==1.5.1
tqdm==4.42.1
numpy==1.18.1
seaborn==0.10.0
nilearn==0.8.0
pandas==1.0.5
pytorch_nlp==0.5.0
opencv_python==4.5.1.48
cmasher==1.3.5
easydict==1.9
imageio==2.8.0
Pillow==8.3.2
scikit_learn==0.24.2
scikit_image
statannot==0.2.3
tensorboardX==2.4